requirements.txt
absl-py==2.1.0
altair==5.4.1
annotated-types==0.7.0
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.5.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
config==0.5.1
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
executing==2.0.1
filelock==3.15.4
Flask==3.0.0
Flask-WTF==1.2.1
flatbuffers==24.3.25
fonttools==4.53.0
frozendict==2.4.4
fsspec==2024.6.1
gast==0.5.4
generativeai==0.0.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.2
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
h5py==3.11.0
html5lib==1.1
httplib2==0.22.0
huggingface-hub==0.25.1
idna==3.7
ipython==8.25.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
loginform==1.2.0
lxml==4.9.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multitasking==0.0.11
MyApplication==0.1.0
namex==0.0.8
narwhals==1.9.1
networkx==3.3
numpy==1.26.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
peewee==3.17.6
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.46
proto-plus==1.24.0
protobuf==5.28.2
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
repoze.lru==0.7
requests==2.32.3
rich==13.7.1
Routes==2.5.1
rpds-py==0.20.0
rsa==4.9
scikit-learn==1.5.0
scipy==1.13.1
sequential==1.0.0
setuptools==70.0.0
six==1.16.0
smmap==5.0.1
soupsieve==2.6
stack-data==0.6.3
streamlit==1.39.0
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
threadpoolctl==3.5.0
toml==0.10.2
torch==2.4.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
watchdog==5.0.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
WTForms==3.1.1
